Separa classe Item() da classe Pessoa()
